automatically backed up by learn

+++ db/migrate/001_create_costumes.rb
@@ -12,10 +12,9 @@ class CreateCostumes < ActiveRecord::Migration[4.2]
   def change
     create_table :costumes do |t|
       t.string :name
-      t.float :price
+      t.integer :price
       t.string :image_url
       t.string :size
-      # t.timestamps null: false
     end
   end
 end
